backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
openai==0.28.0
pandas==2.0.0